scalar  product patched

+++ inc/miniRT.h
@@ -6,7 +6,7 @@
 /*   By: cpapot <cpapot@student.42lyon.fr >         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/09 12:03:42 by cpapot            #+#    #+#             */
-/*   Updated: 2023/06/12 17:10:40 by cpapot           ###   ########.fr       */
+/*   Updated: 2023/06/12 21:06:06 by cpapot           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@
 /**/
 # include <stdio.h>
 
-int	sphere_hited(t_ray camray, t_sphere sphere);
+double	sphere_hited(t_ray camray, t_sphere sphere);
 double	ft_find_light(t_vec_3 normal, t_light light);
 
 /**/
